# :sparkles: enhancement: use msgp for flash message encoding/decoding