--- a/middleware/cache/manager.go
+++ b/middleware/cache/manager.go
@@ -10,7 +10,7 @@ import (
 
 // msgp -file="manager.go" -o="manager_msgp.go" -tests=true -unexported
 //
-//go:generate msgp
+//go:generate msgp -o=manager_msgp.go -tests=true -unexported
 type item struct {
 	headers   map[string][]byte
 	body      []byte