--- a/middleware/limiter/manager.go
+++ b/middleware/limiter/manager.go
@@ -10,7 +10,7 @@ import (
 
 // msgp -file="manager.go" -o="manager_msgp.go" -tests=false -unexported
 //
-//go:generate msgp
+//go:generate msgp -o=manager_msgp.go -tests=false -unexported
 type item struct {
 	currHits int
 	prevHits int