@efectn 2024-08-21
> @efectn can i merge , or should i wait for the coverage extension ? ![image](https://private-user-images.githubusercontent.com/7063188/359886793-75228736-9ee4-4016-982e-51601c4b66ce.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjQyODI4ODgsIm5iZiI6MTcyNDI4MjU4OCwicGF0aCI6Ii83MDYzMTg4LzM1OTg4Njc5My03NTIyODczNi05ZWU0LTQwMTYtOTgyZS01MTYwMWM0YjY2Y2UucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI0MDgyMSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNDA4MjFUMjMyMzA4WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9YTI3OWM0Yzc0MjA4MGIxMzQ4OWMxODNmZDRjZmI4YTY0Njc3YWEwMGMxNTI0MTlhMmE3Yjg5ZjkyZGU1MDU5NSZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QmYWN0b3JfaWQ9MCZrZXlfaWQ9MCZyZXBvX2lkPTAifQ.SCIvZNNVRC-qdwO9AG4iimy7SqhOnQhUxRyGiNvPTcY)

I will complete it tomorrow