@gaby 2024-08-22
@efectn If you can add the command for generating the `msgp` to the Makefile.