@gaby 2024-08-29
> > @efectn If you can add the command for generating the `msgp` to the Makefile.
> 
> you mean this https://github.com/gofiber/fiber/blob/main/Makefile#L61-L67

Yes